refactor(server): isolate PR feedback persistence
